feat: add spline implementation using ext library
Spline external library is used to generate a path
of next points from the car previous path
(for smoothness) plus a desired Frenet path in
the center of lane 1 as provided by localization.
Target speed is set constant and lower than the limit.
Data from object fusion are not considered yet.